CC-18 #back-testing #COMMENT Add get specific Person functionality in PersonController
This update includes changes in classes such as PersonController, PersonService, and PersonServiceImpl. It introduces a new method in PersonController that uses an authentication and ID to retrieve a specific Person. It also adjusts the location of the PersonForm class from the models package to a new sub-package 'person'. Tests and validation also update to reflect these changes.