CC-12 #COMMENT Implement user invitation and account creation functionality
This commit introduces new functionalities to the application, it adds the ability to send invitation emails to users with a unique link for account creation. New files, including UserService and UserController to handle these functions, and created User token model, repository and related configurations. In addition, it has also added a search query in Person Repository for finding a person by their first and last name.