CC-12 #back-testing #COMMENT Add User account creation and invitation functions
This commit provides the ability to send invitation emails for account creation to users. New functionalities were added to UserService and UserController to manage these operations. A UserCreateToken model, which includes token validation and repository, was also implemented. An additional search query was added in PersonRepository to facilitate user identification by their name.