CC-63 #Front #COMMENT Add tests for AddressServiceImpl methods
Introduced test cases for the methods in AddressServiceImpl. These include creating a new address with a valid form, creating with a null form, and creating with a form linked to a non-existent user. Additionally, an exception case was updated in AddressServiceImpl to throw EntityNotFoundException when a user is not found for a given id.