CC-27 #back-testing #COMMENT Implement create method in DogService and DogController
The DogService and DogController classes have both been enhanced with a create method for creating new dog entities. Corresponding changes have been made in the DogServiceImpl class and DogCreateForm class. Now, information about a new dog can be inputted via a form and is then saved to the dog repository.