CC-29 #back-testing #COMMENT Add dog search functionality
Added the ability to search for dogs by their names. This change includes adding a search method to DogService, a specification builder in SpecificationBuilder, new route in the DogController, and extending the DogRepository to use JpaSpecificationExecutor. We also created DogSearchForm and DogShortDTO to handle the search form data and response data respectively.