CC-32 #COMMENT Add delete method to DogService and DogController
Added a delete method to both the DogService interface and its implementation in DogServiceImpl. Correspondingly, a DeleteMapping has been set up in DogController to bind deletion request with the specified ID of a dog to this new service method. This feature allows us to delete a dog record by its ID.