CC-69 #Back-testing #COMMENT Add 'getOne' method to BreedService
Added a new 'getOne' method to the BreedService interface and its implementation in BreedServiceImpl class. This method will allow to fetch a specific breed based on its id from the repository. Also, created a BreedController and a BreedDTO which provides a public API endpoint to retrieve a breed and convert the breed entity to a DTO.