CC-70 #Back-testing #COMMENTAdd create breed feature
The feature to create a new breed has been added. A void create method has been implemented that takes a BreedForm as input in BreedServiceImpl and BreedController. A new breed is built using the BreedForm and saved to the repository. An exception is thrown if the form is empty. The new feature is only accessible to users with 'ROLE_ADMIN' or 'ROLE_HELPER' roles.