CC-71 #Back-testing #COMMENT Add "getAll" method to BreedService
The BreedService now supports retrieving all breeds via a new "getAll" method. This is reflected in the implemented BreedServiceImpl class, where a call to this method makes a "findAll" request to the breed repository. The BreedController was also updated to expose this new functionality through an appropriate HTTP GET mapping.