CC-73 #Front #COMMENT Add create method to VaccineService and its implementation
A 'create' method was added to the VaccineService interface and implemented in VaccineServiceImpl. This method accepts a VaccineForm object and creates and stores a new Vaccine object in the database. Related changes were also reflected in the corresponding test class. Consequentially, a post mapping was added to the VaccineController to handle new vaccine record creation.