CC-77 #Front #COMMENT Implement update method for vaccines
A method to update vaccine records has been added to the VaccineService interface and implemented in the VaccineServiceImpl. An endpoint to handle this action in the VaccineController has also been added. Appropriate tests for these updates have been incorporated into the VaccineServiceImplTest class. Furthermore, a new model, VaccineUpdateForm, was created to manage update forms.