CC-50 #Back-testing #COMMENT Implement DiagnosticDTO and add retrieve diagnostic method
This commit introduces the DiagnosticDTO to standardize and control data transfer. It also corrects a typo in the Diagnostic.java model. Furthermore, it adds a method to retrieve individual diagnostic records in the DiagnosticService and its implementation, expanding the functionality of the DiagnosticService. Finally, a respective endpoint has been added to DiagnosticController to facilitate this functionality.