Implement Morphology feature for canine companion
This commit introduces Morphology, a new feature for the canine companion application. It adds the Morphology model, repository, and service, as well as the necessary DTO and form for the presentation layer. Morphology represents a dog's physical characteristics and it will be connected to the Dog model. Created a new endpoint in the controller for the manipulation of the Morphology data. Also, unit tests have been added to ensure the feature's proper functioning.